// HeroesVillanos/src/app/models/details.model.ts
export interface Details {
  name: string;
  aliases: string[];
  firstAppearance: string;
  placeOfBirth: string;
  fullName: string;
  occupation: string;
  relatives: string;
  base: string;
  image: string;
  groupAffiliation: string;
  intelligence: string;
  strength: string;
  speed: string;
  durability: string;
  power: string;
  combat: string;
}